### 2024-01-19 #152 @lwasser
hi there @eliotwrobson 👋 i wanted to point out two small items that you could address associated with your package's structure!

1. you have a [shell setup.py file](https://github.com/caleb531/automata/blob/main/setup.py) here that you don't need! It did used to be the case that the setup.py file was needed for editable installs. HOwever that is no longer the case. You can remove it and all should be fine.
2. Setuptools also [installs wheel at runtime](https://github.com/pypa/setuptools/blob/1d4c20eb09a9e60fedd326007f15fad99e77b742/setuptools/build_meta.py#L325) as a requirement. so you can [remove wheel from your pyproject.toml build declaration](https://github.com/caleb531/automata/blob/main/pyproject.toml#L37)

These are small changes but will modernize / clean up your package just a bit! :)
