### 2024-01-19 13e5c1b
fix: Small changes from comment (#201)
* Changes from comment

* Changed workflow
