### 2024-01-31 dd6e686
Add Python 3.12 to CI matrix
Also run build and lint CI jobs to ensure that building and linting
works for older Python versions.
