### 2024-01-31 354c921
Update typing_extensions to fix TypeVar error on Python 3.12
