+++ requirements.txt
@@ -26,4 +26,4 @@ pyparsing==3.0.9
 pyproject_hooks==1.0.0
 tomli==2.0.1
 types-frozendict==2.0.9
-typing_extensions==4.5.0
+typing_extensions==4.9.0
