### 2024-01-31 4008f3a
Upgrade flake8 and its dependencies
In an attempt to fix a flake8.exceptions.FailedToLoadPlugin error on
Python 3.12.
