+++ requirements.txt
@@ -4,9 +4,9 @@ cached_method==0.1.0
 click==8.1.3
 coloraide==1.8.2
 coverage==6.4.4
-flake8==6.0.0
+flake8==7.0.0
 flake8-black==0.3.6
-flake8-isort==6.0.0
+flake8-isort==6.1.1
 Flake8-pyproject==1.2.3
 frozendict==2.3.4
 isort==5.10.1
@@ -18,9 +18,9 @@ nose2==0.12.0
 packaging==23.0
 pathspec==0.11.1
 platformdirs==3.2.0
-pycodestyle==2.10.0
+pycodestyle==2.11.1
 pydot==1.4.2
-pyflakes==3.0.1
+pyflakes==3.2.0
 pygraphviz==1.10
 pyparsing==3.0.9
 pyproject_hooks==1.0.0
