### 2024-01-31 45f3ded
Add Python 3.12 to CI matrix (#204)
* Add Python 3.12 to CI matrix

Also run build and lint CI jobs to ensure that building and linting
works for older Python versions.

* Update typing_extensions to fix TypeVar error on Python 3.12

* Upgrade flake8 and its dependencies

In an attempt to fix a flake8.exceptions.FailedToLoadPlugin error on
Python 3.12.
