### 2024-02-01 6152336
First version of new function
