### 2024-02-01 7f1f784
Cleaned up function calls
