### 2024-02-01 36b672c
Update nfa.py
