### 2024-02-01 b13cddd
perf: Graph search speed fix (#205)
Improves speed of graph reachability subroutines
