### 2024-02-02 2d02dbc
fix: Type was too general (#206)
