+++ automata/base/utils.py
@@ -214,14 +214,11 @@ def pairwise(iterable: Iterable[T], final_none: bool = False) -> Iterable[Tuple[
     return zip(a, b)
 
 
-NodeT = TypeVar("NodeT")
-
-
 def get_reachable_nodes(
     G: nx.DiGraph,
-    sources: Iterable[NodeT],
+    sources: Iterable[Any],
     reversed: bool = False,
-) -> Set[NodeT]:
+) -> Set[Any]:
     """
     Return a set with all descendants (or predecessors if reversed is True)
     of the nodes in sources.
