### 2024-02-11 cc41406
Fix "Variable not allowed in type expression" Pylance errors
