### 2024-02-11 31d480f
Do not exclude any partial branches from coverage report
Even if the branches aren't able to be coverage by their nature, it
makes for a more honest coverage score.
