### 2024-02-11 f4c521b
Do not exclude any partial branches from coverage report (#208)
Even if the branches aren't able to be coverage by their nature, it
makes for a more honest coverage score.
