+++ requirements.txt
@@ -1,20 +1,20 @@
-black==23.3.0
+black==24.2.0
 build==1.0.3
 cached_method==0.1.0
 click==8.1.3
 coloraide==1.8.2
-coverage==6.4.4
+coverage==7.4.1
 flake8==7.0.0
 flake8-black==0.3.6
 flake8-isort==6.1.1
 Flake8-pyproject==1.2.3
 frozendict==2.3.4
-isort==5.10.1
+isort==5.13.2
 mccabe==0.7.0
-mypy==1.1.1
+mypy==1.8.0
 mypy-extensions==1.0.0
 networkx==2.6.2
-nose2==0.12.0
+nose2==0.14.1
 packaging==23.0
 pathspec==0.11.1
 platformdirs==3.2.0
