### 2024-03-03 d071b71
Update dfa.py
