### 2024-03-03 5ccaaf1
Aho-Corasick (#167)
* Aho-Corasick for multiple substring construction

* Add simple tests for Aho-Corasick

* Type annotations

* Fix type hints

* Fix type hints

* Update test_dfa.py

* Update dfa.py

* Updated docstring

* Update dfa.py

* Update dfa.py

* Add annotation

* Reformat docstring

* Update test_dfa.py

---------

Co-authored-by: Eliot Robson <eliot.robson24@gmail.com>
