### 2024-03-12 de38ac2
Add classifiers and update contributing guide
