### 2024-03-13 c8b1148
Add classifiers and update contributing guide (#210)
