### 2024-03-23 e1064fc
Update pygraphviz and document installation fix on macOS
