### 2024-03-23 c93cf63
Downgrade pygraphviz to 1.10
This should fix Pytohn 3.8 and 3.9 compatibility.
