### 2024-03-23 8b5e8ea
Install other dependencies after correctly installing pygraphviz
