+++ CONTRIBUTING.md
@@ -78,6 +78,22 @@ virtualenv is active):
 pip install -r requirements.txt
 ```
 
+#### Troubleshooting pygraphviz
+
+If you run into any trouble building the wheel for `pygraphviz` on macOS when
+installing dependencies, try running:
+
+```sh
+brew install graphviz
+python3 -m pip install -U --no-cache-dir  \
+        --config-settings="--global-option=build_ext" \
+        --config-settings="--global-option=-I$(brew --prefix graphviz)/include/" \
+        --config-settings="--global-option=-L$(brew --prefix graphviz)/lib/" \
+        pygraphviz==1.10
+# Proceed to install other dependencies
+pip install -r requirements.txt
+```
+
 ### Running unit tests
 
 The project's unit tests are written using [unittest][unittest] and run using
