### 2024-03-23 b55b871
Add missing docstring to DFA.from_substrings test
