+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "automata-lib"
-version = "8.2.0"
+version = "8.3.0"
 description = "A Python library for simulating finite automata, pushdown automata, and Turing machines"
 readme = "README.md"
 requires-python = ">=3.8"
