### 2024-03-25 #152 @sneakers-the-rat
Looks great so far @phildong  -

> Am I supposed evaluate the novelty/impact the package?

there has been some ongoing discussion about whether this should matter in pyopensci reviews - at the moment the answer is **no,** don't worry about novelty/impact. The EiC has already cleared the package as in-scope. You are welcome to comment about this in your review, as well as request links to prior work as relevant: putting work in context is an important part of reviewing, and there is one item in the checklist asking about "how the package compares to other similar packages." It is a generally good practice for packages to maintain a "see also" section in their docs if there are other packages that are closely related, along with a note explaining similarities and differences: "Our package is for x, if you are looking for y then consider z package," and so on. You are also welcome to open an issue asking the authors about this if you aren't sure about the relation to prior work!

> I see the "JOSS publication" section was not checked in the original submission, so do I still need to worry about the corresponding section in the review template?

nope :) you can remove those checkboxes from your review.

> So I'm wondering what's the criteria for examples and what's the best way to proceed?

This is a tough judgement call to make. re: API documentation - what counts as a package's public API anyway? A decent heuristic is "the set of functions that one would expect to use to accomplish the stated purpose of the package." So even if they aren't marked private by a leading underscore, you might not require docstrings for every function in a `utils` module if they would only be used in an edge case, or are mostly used internally within the package. Another heuristic is 'do these functions show up in example notebooks' (aside from, well, utility functions that are explicitly for use in notebooks eg. for pretty printing and the like). Ultimately it's up to you as the reviewer whether the package is documented in a way that allows you to evaluate it, and would allow your average bear to use it.
