### 2024-05-15 a9d2dba
Added randomized regex example
