### 2024-05-16 6cfc2ce
Added test with bugfix
