### 2024-05-16 d51edb7
Added check for edge case
