### 2024-05-16 3b029cf
fix: Fix bug in minify edge case (#218)
