+++ automata/fa/dfa.py
@@ -338,6 +338,7 @@ class DFA(fa.FA):
             An equivalent complete DFA.
 
         """
+
         if not self.allow_partial:
             return self.copy()
 
@@ -624,7 +625,9 @@ class DFA(fa.FA):
                                 x for x in count(-1, -1) if x not in reachable_states
                             )
                             for trap_symbol in input_symbols:
-                                transition_back_map[trap_symbol][trap_state] = []
+                                transition_back_map[trap_symbol][trap_state] = [
+                                    trap_state
+                                ]
 
                             reachable_states.add(trap_state)
 
@@ -680,6 +683,11 @@ class DFA(fa.FA):
             for state in eq
             if trap_state not in eq
         }
+
+        # If only one equivalence class with the trap state,
+        # return empty language.
+        if not back_map:
+            return cls.empty_language(input_symbols)
 
         new_input_symbols = input_symbols
         new_states = frozenset(back_map.values())
@@ -1707,13 +1715,18 @@ class DFA(fa.FA):
 
         states = frozenset(transitions.keys())
         final_states = {last_state}
+
+        is_partial = any(
+            len(lookup) != len(input_symbols) for lookup in transitions.values()
+        )
+
         return cls(
             states=states,
             input_symbols=input_symbols,
             transitions=transitions,
             initial_state=0,
             final_states=final_states if contains else states - final_states,
-            allow_partial=as_partial,
+            allow_partial=is_partial,
         )
 
     @classmethod
