### 2024-05-16 5d334af
Add graphviz instructions to README
