### 2024-05-18 #152 @eliotwrobson
Docs PR addressing most of the currently open issues from this review: https://github.com/caleb531/automata/pull/217

Basically adds discussion addressing most of the points of confusion that have come up so far.

Regarding more self-contained tutorials, that's coming up soon 👍👍
