### 2024-06-04 aa0b48b
Update docs/examples/fa-examples.md
Co-authored-by: Caleb Evans <caleb@calebevans.me>
