### 2024-06-04 0956873
Update docs/examples/perf-examples.md
Co-authored-by: Caleb Evans <caleb@calebevans.me>
