### 2024-06-04 6246b26
docs: Additions for review (#217)
* Added the first examples of perf

* Added randomized regex example

* Fixed typos

* Extend examples and add note

* Enhanced examples

* Update docs/examples/fa-examples.md

Co-authored-by: Caleb Evans <caleb@calebevans.me>

* Update docs/examples/perf-examples.md

Co-authored-by: Caleb Evans <caleb@calebevans.me>

---------

Co-authored-by: Caleb Evans <caleb@calebevans.me>
