+++ docs/examples/fa-examples.md
@@ -1,12 +1,27 @@
-# FA Examples
+# Finite Automata Examples
 
 On this page, we give some short examples with discussion for the finite
 automata classes and methods.
 
 ## Reading basic input
 
-The following code snippet creates a `DFA` and prints whether it accepts or
-rejects user input.
+In this example, we define a function that takes in an automaton
+and asks the user for input strings, printing whether the input was
+accepted or rejected:
+
+```python
+def read_user_input(my_automaton):
+    try:
+        while True:
+            if my_automaton.accepts_input(input("Please enter your input: ")):
+                print("Accepted")
+            else:
+                print("Rejected")
+    except KeyboardInterrupt:
+        print("")
+```
+
+With this code in hand, let's define a DFA and see it in action:
 
 ```python
 from automata.fa.dfa import DFA
@@ -24,76 +39,113 @@ my_dfa = DFA(
     final_states={'q1'}
 )
 
-try:
-    while True:
-        if my_dfa.accepts_input(input('Please enter your input: ')):
-            print('Accepted')
-        else:
-            print('Rejected')
-except KeyboardInterrupt:
-    print('')
+read_user_input(my_dfa)
+```
+
+We can also do the same with an NFA we define. Note that the
+transition dictionary for the NFA has a different structure than
+that of the DFA, and that we are working over a different input
+alphabet than the previous example:
+
+```python
+from automata.fa.nfa import NFA
+
+# NFA which matches strings beginning with "a", ending with "a", and
+# containing no consecutive "b"s
+my_nfa = NFA(
+    states={"q0", "q1", "q2"},
+    input_symbols={"a", "b"},
+    transitions={
+        "q0": {"a": {"q1"}},
+        "q1": {"a": {"q1"}, "": {"q2"}},
+        "q2": {"b": {"q0"}},
+    },
+    initial_state="q0",
+    final_states={"q1"},
+)
+
+read_user_input(my_nfa)
 ```
 
 ## Subset for NFAs
 
 The `NFA` does not have a built-in method for checking whether it is a subset
-of another `NFA`. However, this can be done using existing methods.
+of another `NFA`. However, this can be done using existing methods in the
+package. See the following function and example usages:
 
 ```python
-# In the following, we have nfa1 and nfa2 and want to determine whether
-# nfa1 is a subset of nfa2.
+import string
+from automata.fa.nfa import NFA
 
-# If taking the union of nfa2 with nfa1 is equal to nfa2 again,
-# nfa1 didn't accept any strings that nfa2 did not, so it is a subset.
-if (nfa1 | nfa2) == nfa2:
-    print('nfa1 is a subset of nfa2.')
-else:
-    print('nfa1 is not a subset of nfa2.')
+def is_subset(nfa1, nfa2):
+    # In the following, we have nfa1 and nfa2 and want to determine whether
+    # nfa1 is a subset of nfa2.
 
+    # If taking the union of nfa2 with nfa1 is equal to nfa2 again,
+    # nfa1 didn't accept any strings that nfa2 did not, so it is a subset.
+    return nfa1.union(nfa2) == nfa2
+
+
+alphabet = set(string.ascii_lowercase)
+
+nfa1 = NFA.from_regex("abc", input_symbols=alphabet)
+nfa2 = NFA.from_regex("(abc)|(def)", input_symbols=alphabet)
+nfa3 = NFA.from_regex("a*bc", input_symbols=alphabet)
+
+print(is_subset(nfa1, nfa2))  # True
+print(is_subset(nfa1, nfa3))  # True
+print(is_subset(nfa2, nfa3))  # False
 ```
 
 ## Edit distance automaton
 
 The following example is inspired by [this blog post][levelshtein-article].
 Essentially, we want to determine which strings in a given set are within
[…]
