### 2024-06-06 #152 @lwasser
hi there team 👋🏻 I wanted to check in on this review. your fearless editor @sneakers-the-rat is focused on  other more pressing issues now going on in their professional environment. As such i'm just checking in to see if we can support Jonny in moving this review forward while they focus on bigger more important things.

As such, i'm trying to figure out the state of this review!
i see one review submitted from @phildong  -- thank you!!
@irisdyoung i suspect you submitted some issues around the package but i don't see your format review submission using [this template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) (please excuse me if i missed it!! )

it's important to have a tracebable record of the review so we ask that you both submit a formal review as a comment using that template and if you open issues / pr's (amazing!!) please link them back to this issue for traceability!

in the meantime I am also going to look for a pinch hitter editor to support this review to keep things moving forward.

everyone - please let me know how this sounds / if you have questions, concerns, etc / if i missed something!! ✨

thank you all for supporting our peer review process.
