### 2024-06-06 #152 @lwasser
ok great.
1. you are welcome to push back and we can discuss . I think the importance of the fix relative to it's effort involved is more important than how long the review takes. As our goal is really that the package works, is usable etc if that makes sense. But we also want to ensure the reviews aren't overly imposing on an author / maintainer! so let's discuss.
2. we are looking for another editor to help get this review over the finish line
3. it will be important for @irisdyoung to submit a formal review here. because as it stands, when looking at the review it appears to only have one review (while i do understand issues have been opened). we do need that review template filled out so we have a formal record of two reviews! 👐🏻

we will get things wrapped up here @eliotwrobson let's work together on getting all three of the above elements sorted out!
