### 2024-06-07 #152 @lwasser
ahhh ok i hear you @eliotwrobson !

so let me make a small list of todos' for this issue here - Eliot - can you fill in any gaps for me please?  i haven't been able to dig deeply into this review but do want to help it move forward

- [ ] there is one potentially outstanding request from @phildong to add "tutorials" or quick start examples for using automata. btw i really appreciate the issue that you opened [here](https://github.com/caleb531/automata/pull/212)!! BUT with that said i see [this here](https://caleb531.github.io/automata/examples/fa-examples/) . This is what we might consider a quickstart or "vignette" in the r world! so you do have that!

I also see that for someone who doesn't know what automata is, it would be hard to follow that vignette and actually understand what it was doing. I wonder if we could discuss this in our slack with the community - the question here is about how rigorous docs should be when we accept vs if there is something that is good enough for now - and we know the authors are actively working on this. What do you all think?

- [ ] @eliotwrobson have you considered updating your github repo link to be your documentation rather than pypi so it's easier to find for people? At first i thought you didn't have docs at all but you do - and they are nice! (mkdocs is so clean!).
- [ ] we are missing a second review submitted to this issue from @irisdyoung .

*********
i think we can discuss the first check above together. either here / on more synchronously in our slack. let me know what you all prefer. in the meantime i am going to pose the question in slack as it's really a standards related question for our review.
