### 2024-06-07 #152 @eliotwrobson
@lwasser thanks! discussing on slack is probably easier, I'm not on there currently but you can send me an invite at eliot.robson24@gmail.com!
