### 2024-06-10 caleb531/automata#225 @eliotwrobson
Possibly update GitHub Link?
As was pointed out [in this comment](https://github.com/pyOpenSci/software-submission/issues/152#issuecomment-2155293765), it might be a good idea to change the GitHub link from the pypi page to the documentation site. I'm not totally sure what the standard practice is, but I guess it's down to personal preference. What do you think @caleb531? I don't have super strong feelings either way.
