### 2024-06-11 505e1b0
Update fa-examples.md
