### 2024-06-11 1cb6622
Update fa-examples.md
