### 2024-06-11 f95b0d1
Added extra test cases and comment about bug
