### 2024-06-11 d55a878
Partial behavior change and corresponding tests
