### 2024-06-14 caleb531/automata#225 @caleb531
@eliotwrobson Are you referring to the GitHub website link in the About section of the repository? If this i what you're referring to, then I am inclined to agree. The PyPI page really (and also quite literally) just contains the same info as the README, anyway.
