### 2024-06-14 caleb531/automata#225 @caleb531
@eliotwrobson It seems clear now that's what you meant, so I went ahead and updated it! Closing this issue. Please feel free to let pyOpenSci know that we've updated this now.
