### 2024-06-21 70ddad2
Partial behavior change and corresponding tests (#228)
* Partial behavior change and corresponding tests

* lint

* Reorder tests

* mypy
