### 2024-06-22 #152 @NimaSarajpoor
**TL;DR**
**Probably better to ignore (5)**

----

@lwasser

> finally i am curious about the description being requested - i do see this

You are absolutely right that the description was provided in the API docs. My comment was mainly about having the methods in the `doctoring` of the class

https://github.com/caleb531/automata/blob/530016a03703e071b72f1665769770039018536d/automata/fa/dfa.py#L54

@eliotwrobson

> The only one I had questions on was 5. Would this just be listing every public method in the top docstring for the DFA class?

> However, I'm not sure what common practice is in other libraries.

After you ask the question, I realized I was not sure myself! 😅  I did some research and found [this stackoverflow post](https://stackoverflow.com/questions/62444612/should-i-list-class-methods-in-the-class-docstring) in which the answer quotes the following sentence from [PEP 257](https://peps.python.org/pep-0257/):

```
The docstring for a class should summarize its behaviour and list the public methods and instance variables.
```

Still, the author of the answer mentioned that they prefer to avoid doing this because it can become out of date in the future.
