### 2024-06-23 097dbfd
Add output to fa examples
