+++ docs/index.md
@@ -14,7 +14,7 @@ processing large inputs. Visualization logic has also been implemented. This pac
 both researchers wishing to manipulate automata and for instructors teaching courses on theoretical
 computer science.
 
-For an overview on automata theory, see [this wikipedia article][wikipedia-article], and
+For an overview on automata theory, see [this Wikipedia article][wikipedia-article], and
 for a more comprehensive introduction to each of these topics, see [these lecture notes][lecture-notes].
 
 [wikipedia-article]: https://en.wikipedia.org/wiki/Automata_theory
