### 2024-06-23 3a8f033
docs: Add links and more discussion to examples. (#227)
* Added links and more discussion

* Added various links

* Update fa-examples.md

* Update fa-examples.md

* Update links in readme

* Add output to fa examples

* Fix various typos

Co-authored-by: Caleb Evans <caleb@calebevans.me>

---------

Co-authored-by: Caleb Evans <caleb@calebevans.me>
