+++ docs/examples/fa-examples.md
@@ -1,11 +1,19 @@
 # Finite Automata Examples
 
 On this page, we give some short examples with discussion for the finite
-automata classes and methods.
+automata (sometimes called finite state machines) classes and methods in
+this package. At a high level, a finite automaton (FA) is an abstract
+machine that can be in any one of a finite number of _states_, and
+moves between states based on a _transition function_ in response to
+reading characters from an input string. The FA will _accept_ or _reject_ an
+input string depending on its current state.
 
-## Reading basic input
+For a detailed overview of this topic, see [this Wikipedia article][wikipedia-fsm]
+or [these lecture notes][lecture-notes].
 
-In this example, we define a function that takes in an automaton
+## Reading input
+
+In this example, we first define a function that takes in an automaton
 and asks the user for input strings, printing whether the input was
 accepted or rejected:
 
@@ -21,7 +29,10 @@ def read_user_input(my_automaton):
         print("")
 ```
 
-With this code in hand, let's define a DFA and see it in action:
+### Deterministic finite automaton (DFA)
+
+To use this function, let's first define a DFA.
+For a detailed definiton, see [this Wikipedia article on DFAs][wikipedia-dfa].
 
 ```python
 from automata.fa.dfa import DFA
@@ -38,14 +49,33 @@ my_dfa = DFA(
     initial_state='q0',
     final_states={'q1'}
 )
-
-read_user_input(my_dfa)
 ```
+
+We can generate a picture of our DFA using the package:
+
+```python
+my_dfa.show_diagram()
+```
+
+This produces the following:
+
+![my dfa image](img/my_dfa.svg)
+
+Now that we've defined our DFA, we can see our funciton in action:
+
+```python
+read_user_input(my_dfa)
+# 001 -> Accepted
+# 011 -> Rejected
+# 000111 -> Accepted
+```
+
+### Nondeterministic finite automaton (NFA)
 
 We can also do the same with an NFA we define. Note that the
 transition dictionary for the NFA has a different structure than
 that of the DFA, and that we are working over a different input
-alphabet than the previous example:
+alphabet than the previous example. For a detailed definiton, see [this Wikipedia article on NFAs][wikipedia-nfa].
 
 ```python
 from automata.fa.nfa import NFA
@@ -63,15 +93,32 @@ my_nfa = NFA(
     initial_state="q0",
     final_states={"q1"},
 )
+```
 
+Similar to the DFA, we can generate a picture of our NFA:
+
+```python
+my_nfa.show_diagram()
+```
+
+This produces the following:
+
+![my nfa image](img/my_nfa.svg)
+
+We can call our function as in the prior example:
+
+```python
 read_user_input(my_nfa)
+# b -> Rejected
+# aa -> Accepted
+# abaa -> Accepted
 ```
 
 ## Subset for NFAs
 
 The `NFA` does not have a built-in method for checking whether it is a subset
 of another `NFA`. However, this can be done using existing methods in the
-package. See the following function and example usages:
+package:
 
 ```python
 import string
@@ -84,14 +131,24 @@ def is_subset(nfa1, nfa2):
     # If taking the union of nfa2 with nfa1 is equal to nfa2 again,
     # nfa1 didn't accept any strings that nfa2 did not, so it is a subset.
     return nfa1.union(nfa2) == nfa2
+```
 
+To see our function in action, we need to define some NFAs. We can
+do this easily by converting from regular expressions. For more information
[…]
