### 2024-06-23 #152 @eliotwrobson
@lwasser thanks for checking in! We just merged https://github.com/caleb531/automata/pull/227, which finishes addressing @/NimaSarajpoor's comments and the last open PR from this review (closing because our examples are now more extensive, and more comprehensive tutorials will be coming later). So I think we've addressed every open issue that's been raised thus far.
