### 2024-06-24 #152 @lwasser
ok y'all. It seems to me that in the latest PR all of the documentation issues have been addressed. We have one review formally submitted to this review from @phildong  - thank you!! @phildong do you have content with the state of the package now after your review? Please just reply here if you have a moment.

@NimaSarajpoor i think from your statement above that you are ok with the state of the package docs  (which was your focus here).
can you please confirm by responding here?

NOTE: I did start a conversation on slack around that PEP. That seems fairly extensive - especially given we have tools to generate API docs and that pep was from 2001. BUT there could be other reasons for it too that I am missing (such as for those running `help()`). regardless if you can confirm that the changes made seem adequate that would be great.

@irisdyoung i just hope things are ok with you. I certainly understand hectic lives. I hope you are well! Please know that we never want to pressure / stress reviewers and volunteers at pyOpenSci and only wish for the best for everyone! Sometimes life gets busy and that is ok / (in fact recommended!!) to take care of yourself first!
