### 2024-06-27 b729cd1
Fix infinite loop bug when length is limited.
