### 2024-06-27 cfeac6b
Re-enable disabled test
