### 2024-06-27 8aa9e79
Add an edge case test for predecessor
