### 2024-06-27 6b272f8
Successors length limits (#224)
* Succesor length limits

* Added extra test cases and comment about bug

* Fix infinite loop bug when length is limited.

* Re-enable disabled test

* Add an edge case test for predecessor

* Formatting

---------

Co-authored-by: Eliot Robson <eliot.robson24@gmail.com>
