### 2024-06-28 ce949fb
Revert "Successors length limits (#224)" (#229)
This reverts commit 39ca92a1bf738c4e8061aea8b849c62c7f73bb0b.
