+++ automata/fa/dfa.py
@@ -1276,8 +1276,6 @@ class DFA(fa.FA):
         *,
         strict: bool = True,
         key: Optional[Callable[[Any], Any]] = None,
-        min_length: int = 0,
-        max_length: Optional[int] = None,
     ) -> Optional[str]:
         """
         Returns the first string accepted by the DFA that comes before
@@ -1293,10 +1291,6 @@ class DFA(fa.FA):
         key : Optional[Callable], default: None
             Function for defining custom lexicographical ordering. Defaults to using
             the standard string ordering.
-        min_length : int, default: 0
-            Limits generation to words with at least the given length.
-        max_length : Optional[int], default: None
-            Limits generation to words with at most the given length.
 
         Returns
         ------
@@ -1309,13 +1303,7 @@ class DFA(fa.FA):
             Raised if the language accepted by self is infinite, as we cannot
             generate predecessors in this case.
         """
-        for word in self.predecessors(
-            input_str,
-            strict=strict,
-            key=key,
-            min_length=min_length,
-            max_length=max_length,
-        ):
+        for word in self.predecessors(input_str, strict=strict, key=key):
             return word
         return None
 
@@ -1325,8 +1313,6 @@ class DFA(fa.FA):
         *,
         strict: bool = True,
         key: Optional[Callable[[Any], Any]] = None,
-        min_length: int = 0,
-        max_length: Optional[int] = None,
     ) -> Generator[str, None, None]:
         """
         Generates all strings that come before the input string
@@ -1342,10 +1328,6 @@ class DFA(fa.FA):
         key : Optional[Callable], default: None
             Function for defining custom lexicographical ordering. Defaults to using
             the standard string ordering.
-        min_length : int, default: 0
-            Limits generation to words with at least the given length.
-        max_length : Optional[int], default: None
-            Limits generation to words with at most the given length.
 
         Returns
         ------
@@ -1359,14 +1341,7 @@ class DFA(fa.FA):
             Raised if the language accepted by self is infinite, as we cannot
             generate predecessors in this case.
         """
-        yield from self.successors(
-            input_str,
-            strict=strict,
-            reverse=True,
-            key=key,
-            min_length=min_length,
-            max_length=max_length,
-        )
+        yield from self.successors(input_str, strict=strict, reverse=True, key=key)
 
     def successor(
         self,
@@ -1374,8 +1349,6 @@ class DFA(fa.FA):
         *,
         strict: bool = True,
         key: Optional[Callable[[Any], Any]] = None,
-        min_length: int = 0,
-        max_length: Optional[int] = None,
     ) -> Optional[str]:
         """
         Returns the first string accepted by the DFA that comes after
@@ -1391,23 +1364,13 @@ class DFA(fa.FA):
         key : Optional[Callable], default: None
             Function for defining custom lexicographical ordering. Defaults to using
             the standard string ordering.
-        min_length : int, default: 0
-            Limits generation to words with at least the given length.
-        max_length : Optional[int], default: None
-            Limits generation to words with at most the given length.
 
         Returns
         ------
         str
             The first string accepted by the DFA lexicographically before input_string.
         """
-        for word in self.successors(
-            input_str,
-            strict=strict,
-            key=key,
-            min_length=min_length,
-            max_length=max_length,
-        ):
+        for word in self.successors(input_str, strict=strict, key=key):
             return word
         return None
 
@@ -1418,8 +1381,6 @@ class DFA(fa.FA):
         strict: bool = True,
         key: Optional[Callable[[Any], Any]] = None,
         reverse: bool = False,
-        min_length: int = 0,
-        max_length: Optional[int] = None,
     ) -> Generator[str, None, None]:
         """
         Generates all strings that come after the input string
@@ -1437,10 +1398,6 @@ class DFA(fa.FA):
[…]
