+++ tests/test_dfa.py
@@ -1836,9 +1836,6 @@ class TestDFA(test_fa.TestFA):
         actual = list(dfa.predecessors("010", strict=False))
 
         self.assertEqual(dfa.predecessor("000"), "00")
-        self.assertEqual(dfa.predecessor("000", max_length=1), "0")
-        self.assertEqual(dfa.predecessor("0", min_length=2), None)
-        self.assertEqual(dfa.predecessor("0000", min_length=2, max_length=3), "000")
         self.assertEqual(dfa.predecessor("0100"), "010")
         self.assertEqual(dfa.predecessor("1"), "010101111111101011010100")
         self.assertEqual(
@@ -1875,10 +1872,6 @@ class TestDFA(test_fa.TestFA):
         self.assertIsNone(dfa.successor("110"))
         self.assertIsNone(dfa.successor("111111110101011"))
 
-        self.assertEqual(dfa.successor("", min_length=3), "000")
-        self.assertEqual(dfa.successor("", min_length=4), "010101111111101011010100")
-        self.assertEqual(dfa.successor("010", max_length=6), "100")
-
         infinite_dfa = DFA.from_nfa(NFA.from_regex("0*1*"))
         self.assertEqual(infinite_dfa.successor(""), "0")
         self.assertEqual(infinite_dfa.successor("0"), "00")
@@ -1889,10 +1882,6 @@ class TestDFA(test_fa.TestFA):
         self.assertEqual(infinite_dfa.successor("1"), "11")
         self.assertEqual(infinite_dfa.successor(100 * "0"), 101 * "0")
         self.assertEqual(infinite_dfa.successor(100 * "1"), 101 * "1")
-        self.assertEqual(infinite_dfa.successor("", min_length=5), "00000")
-        self.assertEqual(infinite_dfa.successor("000", min_length=5), "00000")
-        self.assertEqual(infinite_dfa.successor("1", min_length=5), "11111")
-        self.assertEqual(infinite_dfa.successor("1111", max_length=4), None)
 
     @params(True, False)
     def test_successor_and_predecessor(self, as_partial: bool) -> None:
