### 2024-06-28 #152 @eliotwrobson
@sneakers-the-rat thank you! We've already set up automatic archiving through zenodo already: https://zenodo.org/records/10864492

I think it would be worthwhile to make a blog post as well, since we can just use examples / material from the JOSS paper or existing examples.

The revisions from items here will be included on Zenodo once we do another release. @caleb531 I've added the pyOpenSci badge to the develop branch along with the other changes, so we should be good to go. Would you like me to write the release notes? (can follow-up on slack if more discussion is needed).
