### 2024-06-28 #152 @sneakers-the-rat
great! added the DOI at the top, let me know when you make that release and that should be all!
