### 2024-06-29 #152 @caleb531
@sneakers-the-rat Perfect! I've just published Automata v8.4.0 with all of the discussed changes and improvements!

https://github.com/caleb531/automata/releases/tag/v8.4.0
https://pypi.org/project/automata-lib/8.4.0/
https://zenodo.org/records/12594180

cc @eliotwrobson
