### 2024-06-30 #152 @sneakers-the-rat
yes, also asked about that & how to format/submit/etc. I would assume it would be a jekyll-markdown style post like those in https://github.com/pyOpenSci/pyopensci.github.io/tree/main/_posts

when i get further info i'll let you know (and also update our docs, bc they are unclear to authors and editors alike at the moment ;)
